Elo   | -3.34 +- 4.57 (95%)
SPRT  | 8.0+0.08s Threads=1 Hash=8MB
LLR   | -0.89 (-2.94, 2.94) [-8.00, 0.00]
Games | N: 7178 W: 1757 L: 1826 D: 3595
Penta | [87, 853, 1770, 800, 79]
https://openbench.tstawowy.pl/test/58/

bench: 2837321

+++ src/movepicker.rs
@@ -180,7 +180,7 @@ impl MovePicker {
         }
 
         if search_state.ply < 256 {
-            if search_state.killer_moves[search_state.ply] == mv {
+            if search_state.killer_moves[search_state.ply as usize] == mv {
                 return FIRST_KILLER_BONUS;
             }
         }
